Fixed `download-from-firehose` tool and added metadata to Firehose response
To fix the tool once and for all, we udated Firehose interface to return extra metadata about the block being served. This metadata is essentially a view on most of `pbbstream.Block` field, just wrap in a conveninent way for transport.

The tool `download-from-firehose` has been made to work with updated servers that serves the `Metadata` field in their response. In the event the server is not recent enough to support `Metadata`, fallback to prior logic is perform so this is backward compatible.